Fix idempotence for wg-portal role
Delegate the /etc/wireguard directory creation to the package, this way,
if later the wg-portal role was to be installed and update that
directory ACL, it won't fail idempotent test.

Ignore the wireguard configuration creation, for some reason the usage
of dictionary is not preventing the detection of when items change!