Revert adding user.js config
Since the configs were mainly to control UI pref (which is not usable
with user.js, as it will always override manual customization)
The rest of the features provided by user.js isn't necessary and are
already covered by policies.

So for now, I'm removing this user.js integration, maybe revisit it
later if it's proven to be useful, ATM, it complicates and brittles the
setup without providing any additional value.